requirements.txt
paho-mqtt>=1.5.1
cryptography>=3.4.6